use font in layout instead
