+++ naurffxiv/src/app/globals.css
@@ -1,5 +1,3 @@
-@import url('https://fonts.googleapis.com/css2?family=Roboto:wght@400;700&display=swap');
-
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
@@ -19,7 +17,6 @@
 body {
   color: var(--foreground-hex);
   background: var(--background-hex);
-  font-family: 'Roboto', sans-serif;
 }
 
 @layer utilities {
