+++ naurffxiv/src/app/theme.js
@@ -1,27 +1,4 @@
 import { createTheme } from "@mui/material/styles";
-import { Roboto } from "next/font/google";
-
-const roboto = Roboto({
-  weight: ['300','400','500','700'],
-  style: ['normal','italic'],
-  subsets: ['latin'],
-  display: 'swap',
-})
-
-const sharedFont = {
-  fontFamily: [
-    roboto.style.fontFamily,
-    '"Segoe UI"',
-    '-apple-system',
-    'BlinkMacSystemFont',
-    '"Helvetica Neue"',
-    'Arial',
-    'sans-serif',
-    '"Apple Color Emoji"',
-    '"Segoe UI Emoji"',
-    '"Segoe UI Symbol"',
-  ].join(','),
-}
 
 // Shared components styling that won't change with theme
 const sharedComponents = {
@@ -55,9 +32,6 @@ const sharedComponents = {
 };
 
 const darkTheme = createTheme({
-  typography: {
-    ...sharedFont
-  },
   palette: {
     mode: "dark",
     background: {
@@ -78,9 +52,6 @@ const darkTheme = createTheme({
 });
 
 const lightTheme = createTheme({
-  typography: {
-    ...sharedFont
-  },
   palette: {
     mode: "light",
     primary: {
